usernames in files _ better script

+++ dockerSandboxes/makecontainerandrun.sh
@@ -5,6 +5,7 @@
 
 FILENAME="$1"
 FILEFOLDER="$2"
+ACTIVITY="$3"
 CNAME=$FILENAME
 
 eval "$(docker-machine env default)"
@@ -12,11 +13,18 @@ eval "$(docker-machine env default)"
 docker run -d -it --name $CNAME umutcoderunner/haskell bash > /dev/null
 docker cp $FILEFOLDER/hsfiles/$FILENAME.hs $CNAME:/home/umutcoderunner/
 docker exec $CNAME runghc /home/umutcoderunner/$FILENAME.hs $FILENAME
-docker exec $CNAME mogrify -format jpeg /home/umutcoderunner/$FILENAME.bmp
-docker cp $CNAME:/home/umutcoderunner/$FILENAME.jpeg $FILEFOLDER/outputimages/
-docker exec $CNAME rm /home/umutcoderunner/$FILENAME.bmp
+
+#special for images
+if [ "$ACTIVITY" == "images" ] 
+then
+  docker exec $CNAME mogrify -format jpeg /home/umutcoderunner/$FILENAME.bmp
+  docker cp $CNAME:/home/umutcoderunner/$FILENAME.jpeg $FILEFOLDER/outputimages/
+  docker exec $CNAME rm /home/umutcoderunner/$FILENAME.bmp
+fi
+
 docker stop $CNAME > /dev/null
 docker rm $CNAME > /dev/null
 
+rm $FILEFOLDER/hsfiles/$FILENAME.hs
